Move `readAll()` to internals dir
