+++ src/impl/index.ts
@@ -1,1 +1,2 @@
 export * from './add-response-header';
+export * from './read-all';
