+++ src/router/dispatch/dispatch-by-name.handler.spec.ts
@@ -1,6 +1,7 @@
 import { RequestContext } from '../../core';
 import { httpListener, Rendering, RenderMeans } from '../../http';
-import { readAll, suppressedLog, testServer, TestServer } from '../../spec';
+import { readAll } from '../../impl';
+import { suppressedLog, testServer, TestServer } from '../../spec';
 import { RouterMeans } from '../router.means';
 import { Routing } from '../routing.capability';
 import { dispatchByName } from './dispatch-by-name.handler';
