+++ src/spec/index.ts
@@ -1,3 +1,2 @@
-export * from './read-all';
 export * from './suppressed-log';
 export * from './test-server';
